Refactoring Classes
