+++ Lab5/src/mineSweeperObjects/Flag.java
@@ -1,0 +1,5 @@
+package mineSweeperObjects;
+
+public class Flag extends MineSweeperObject{
+
+}
