+++ Lab5/src/mineSweeperObjects/Mine.java
@@ -1,0 +1,5 @@
+package mineSweeperObjects;
+
+public class Mine extends MineSweeperObject {
+
+}
