+++ Lab5/src/mineSweeperObjects/ProximityNumber.java
@@ -1,0 +1,8 @@
+package mineSweeperObjects;
+
+public class ProximityNumber extends MineSweeperObject {
+	private int proxNumber;
+	private String proxString;
+	
+
+}
